move up generate utils
