+++ src/generate/index.ts
@@ -1,3 +1,3 @@
 export * from './AbstractGenerate.js'
 export * from './r3f/GenerateR3F.js'
-export * from './r3f/utils.js'
+export * from './utils.js'
